Adding new input handlers
feat: the inputs on Contact.jsx and SignIn.jsx now have responsive and cleaner "onChange" handlers

fix: [Contact.jsx] the "Send" button is now being displayed correctly

It had a responsiveness bug and it was entering inside of the Footer when displayed in lower resolutions. I fixed it by increasing the section "pb" class from "pb-20" to "pb-72"

style: [Contact.jsx] changed the width class of "Send" button to "full"

A button with the same size of the other inputs looks better and it's more modern than a default-sized button.